requirements.txt
jflask==2.3.3
PyMuPDF==1.22.5
pytesseract==0.3.10
pillow==10.0.0
pandas==2.0.3
openai==0.28.1
python-dotenv==1.0.0
opencv-python==4.8.0.76
numpy==1.24.3
